Update convert_dataframe.py
Some minute changes based on discussions with @CalCraven. I added the "molecule.name" to be a default column for the "sites" parameter. Another update, is a check to see whether a column option is available for the parameter. If it is unavailable, the column is skipped and a warning is printed. (Now that I think about it though, probably the warnings.warn is unnecessary and a simple "print" should do it). A command like this "dfDict = to_dataframeDict(ptop, parameters=["sites", "bonds"], columns = ['molecule.name', 'molecule'], format ='default')" should work now. Let me know if this works for you or if I should try to change it in another way.